+++ UCR/Views/Controls/Plugin/PluginControl.xaml.cs
@@ -1,6 +1,7 @@
 ﻿using System.Windows;
 using System.Windows.Controls;
 using HidWizards.UCR.ViewModels.ProfileViewModels;
+using MaterialDesignThemes.Wpf;
 
 namespace HidWizards.UCR.Views.Controls.Plugin
 {
@@ -22,6 +23,18 @@ namespace HidWizards.UCR.Views.Controls.Plugin
         {
             var viewModel = DataContext as PluginViewModel;
             viewModel?.AddFilter();
+        }
+
+        private void Filter_OnClick(object sender, RoutedEventArgs e)
+        {
+            var filter = ((Chip)sender).DataContext as FilterViewModel;
+            filter?.ToggleFilter();
+        }
+
+        private void Filter_OnDeleteClick(object sender, RoutedEventArgs e)
+        {
+            var filter = ((Chip)sender).DataContext as FilterViewModel;
+            filter?.RemoveFilter();
         }
     }
 }